Refactoring snuba backend into base class to be extended later
